requirements.txt
requests
beautifulsoup4
python-dotenv
pyyaml
playwright
psycopg[binary]
